Add more smarts into apply_branch()
When there is a merge conflict allow the user to fix it in an interactive shell.  Once the conflict is resolved, take over again and automat the commit.

Then when more changes are made locally, know to re-apply the conflict merge and apply new changes atop that to carry over the conflict resolution